Move notifications permission dialog to new pkg
