+++ app/src/main/res/layout/dialog_notifications_permission.xml
@@ -9,7 +9,7 @@
         android:id="@+id/close_button"
         style="@style/CryptoX_ImageButton"
         android:layout_margin="8dp"
-        android:contentDescription="@string/welcome_notification_permission_deny"
+        android:contentDescription="@string/notifications_permission_deny"
         android:padding="8dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -37,7 +37,7 @@
         android:layout_marginHorizontal="24dp"
         android:layout_marginTop="16dp"
         android:gravity="center"
-        android:text="@string/welcome_notification_permission_title"
+        android:text="@string/notifications_permission_title"
         android:textAppearance="@style/CCX_Typography_H2"
         android:textColor="@color/ccx_neutral_tint_7"
         app:layout_constraintStart_toStartOf="parent"
@@ -50,7 +50,7 @@
         android:layout_marginHorizontal="24dp"
         android:layout_marginTop="8dp"
         android:gravity="center"
-        android:text="@string/welcome_notification_permission_details"
+        android:text="@string/notifications_permission_details"
         android:textAppearance="@style/CCX_Typography_Body"
         android:textColor="@color/ccx_neutral_tint_7"
         app:layout_constraintStart_toStartOf="parent"
@@ -60,7 +60,7 @@
         android:id="@+id/allow_button"
         style="@style/CCX_Button_Primary.Modal"
         android:layout_marginTop="16dp"
-        android:text="@string/welcome_notification_permission_allow"
+        android:text="@string/notifications_permission_allow"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/details_text_view" />
@@ -71,7 +71,7 @@
         android:layout_marginTop="6dp"
         android:layout_marginBottom="14dp"
         android:background="@null"
-        android:text="@string/welcome_notification_permission_deny"
+        android:text="@string/notifications_permission_deny"
         android:textColor="@color/ccx_neutral_tint_7"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
